requirements.txt
beautifulsoup4==4.11.1
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
colorama==0.4.6
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
einops==0.2.0
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.25.0
gdown==4.4.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
Jinja2==3.1.2
kiwisolver @ file:///C:/ci/kiwisolver_1653292407425/work
lmdb==1.3.0
MarkupSafe==2.1.1
matplotlib @ file:///C:/b/abs_ae02atcfur/croot/matplotlib-suite_1667356722968/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
munkres==1.1.4
networkx==2.8.8
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy @ file:///C:/b/abs_5ct9ex77k9/croot/numpy_and_numpy_base_1668593740598/work
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/b/abs_97pwvlq0kg/croot/pandas_1668631294821/work
Pillow==9.3.0
ply==3.11
prefetch-generator==1.0.1
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
pypdfium2==3.10.0
PyQt5==5.15.7
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
Shapely==1.8.5.post1
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
tifffile==2022.10.10
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.11.0
torchvision==0.12.0
tornado @ file:///C:/ci/tornado_1662458743919/work
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
vietocr==0.3.11
wincertstore==0.2
